Fix location of Version type
