+++ cmd/agentctl/client/api.go
@@ -12,7 +12,6 @@ import (
 	"go.ligato.io/vpp-agent/v3/client"
 	"go.ligato.io/vpp-agent/v3/cmd/agentctl/api/types"
 	"go.ligato.io/vpp-agent/v3/plugins/kvscheduler/api"
-	restapi "go.ligato.io/vpp-agent/v3/plugins/restapi/types"
 	"go.ligato.io/vpp-agent/v3/proto/ligato/kvscheduler"
 )
 
@@ -31,7 +30,7 @@ type APIClient interface {
 	KVDBClient() (KVDBAPIClient, error)
 	GRPCConn() (*grpc.ClientConn, error)
 	HTTPClient() *http.Client
-	AgentVersion(ctx context.Context) (*restapi.Version, error)
+	AgentVersion(ctx context.Context) (*types.Version, error)
 	NegotiateAPIVersion(ctx context.Context)
 	NegotiateAPIVersionPing(types.Ping)
 	Close() error
