--- plugins/restapi/types/types.go
@@ -1,30 +1,0 @@
-//  Copyright (c) 2020 Cisco and/or its affiliates.
-//
-//  Licensed under the Apache License, Version 2.0 (the "License");
-//  you may not use this file except in compliance with the License.
-//  You may obtain a copy of the License at:
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-//  Unless required by applicable law or agreed to in writing, software
-//  distributed under the License is distributed on an "AS IS" BASIS,
-//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//  See the License for the specific language governing permissions and
-//  limitations under the License.
-
-package types
-
-// Version contains response of Agent REST API:
-// GET "/info/version"
-type Version struct {
-	App       string
-	Version   string
-	GitCommit string
-	GitBranch string
-	BuildUser string
-	BuildHost string
-	BuildTime int64
-	GoVersion string
-	OS        string
-	Arch      string
-}
